requirements.txt
asgiref==3.6.0
click==8.1.7
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.8
Django==4.1.5
django-adminlte3==0.1.6
django-import-export==3.2.0
et-xmlfile==1.1.0
filelock==3.13.3
gunicorn==21.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.16.0
hyperframe==6.0.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
packaging==24.0
platformdirs==4.2.0
priority==2.0.0
PyYAML==6.0.1
sqlparse==0.4.3
tablib==3.5.0
tzdata==2022.7
uvicorn==0.29.0
virtualenv==20.25.1
whitenoise==6.6.0
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
